Extracted Function and Equation classes (Calculator)
